Add ucsc command line utilities
Add ucsc command line utilities, including blat, bigwigtools etc.
Note that there was a problem in setting up the build in that despite
setting SSLDIR to $PREFIX/include/openssl, the openssl header files
could not be found. A workaround was to copy openssl header files to a
directory that was included in compilation commands, namely
kent/src/inc/.