allow crossmap to fail gracefully on py35
The following py2-only print statement is in crossmap's setup.py:

```
Traceback (most recent call last):
File "setup.py", line 11, in <module>
    print >> sys.stderr, "ERROR: CrossMap requires Python 2.7"
```

This raises TypeError on py35 when trying to build the package. Somehow
this was resulting in None as output from the `check_call()`. Anyway,
this allows the fail to happen gracefully on py35, and it builds fine on
py27.