use requests module to be able to track redirects
This ended up being important for creating data package dependencies -
data packages have a different url pattern on the bioconductor site. Now
I'm using the "short URL" for any package (which is listed in the
details table, but has a predictable pattern). That redirects to the
actual bioconductor page for the package. That final destination url can
be used to construct the tarball url.