add note to README.md to mention how to enter a bash shell within the build container
add note to README.md to mention how to enter a bash shell within the build container

Sometimes it is helpful to debug build issues within the docker container by manually calling conda build within a bash shell for the container. This PR adds a note to the readme instructing how to do this, since the default container entrypoint is overridden by the bioconda build script.