invocation fix for Trinity
Previously, Trinity failed to execute downstream when it attempted to
pull in its dependencies. This is because its dependencies were
included from a relative location that was incorrect because Trinity
was being called from an symlink that Trinity/perl did not resolve.
This commit preserves the Trinity symlink, which in turn calls a shell
script that calls Trinity—hopefully solving the problem