Change Perl build to use Conda gcc
The gcc 4.8.5 via the devtoolset-2 RPMs used previously to build Perl
has backports from RH from the GCC 4.9 series.  This includes the
`-fstack-protector-strong` option introduced in 4.9.  The Perl `Configure`
sees this option as supported and adds it to the default compiler options.
When building Perl modules with shared libraries, specifying `gcc` in the
build requirements pulls in the gcc from the default channel.  The build
then fails as this is an unpatched version and that option is not
supported. Using the conda gcc, `Configure` uses `-fstack-protector`
instead and avoids the issue.