fix logic in selecting java version
Previously java would be set if JAVA_HOME was UNset, which I assume was not the intended behaviour. This would cause picard to fail if the system installation was not the correct version to use. 

For me, picard 2.3.0 used `/bin/java` which was version 1.7.x instead of the version installed by conda.