add (installer for) gatk-protected (full, license-restricted GATK)
The license for the full version of GATK prevents it from being
redistributed. This adds a recipe for a GATK installer that provides
the commands “gatk” and “gatk-register”. The former is a thin wrapper
around the gatk.jar file, and the latter copies a user-supplied gatk
jar file (or *.tar.bz file) to the conda environment.