add badge generator
This is a static HTML page derived from one created by @ewels in #1971.
It is capable of rendering SVG and PNG badges with custom text,
suitable for promoting bioconda packages in posters and presentations.
This commit differs from the original in that the badges were
re-created in Inkscape so the SVGs are accessible to those without
Illustrator. The svg files are now loaded in the page dynamically via
javascript so the badge files can be edited externally, obviating the
need to copy+paste from the inline svg source. Inkscape will also
preserve tspans, so it is possible to edit and see design changes
immediately. The modifications also add javascript to download a
base64-encoded data string representation of the SVG designs, as well
as canvas-rendered PNG files at 4x resolution. The “console” text has a
genetic font monospace font family now instead of only Monaco. The SVG
files scale correctly in Inkscape (despite fixed-font for the console
text). Added a link to the docs, with the badge generator files stored
in _static/