add biopython 1.68
At the end of September the NCBI will be deprecating their HTTP Entrez
API in favor of the HTTPS version. A newer version of biopython that
uses the HTTPS version is available (released August 25, 2016), but it
is not yet in the default conda channel. Until it is, we should
probably have a package available.