Update ViennaRNA: build Python 2 and 3 bindings (#3021)
building both interfaces in the same package causes build deps to two python versions. This is solved using a trick, installing the second version in a environment during the build, since easier solutions seem to be precluded by the current build system (no support for defining build deps to different versions of the same package in meta.yml).