peptide-shaker Added --exec_dir option
The --exec_dir option specifies an alternate directory for the
peptide-shaker distribution.  If the option is used and the directory
does not exist, the script will copy the distribution to that directory.
PeptideShaker updates files relative to the path of the jar file. Thus
we
need to copy the jar file, lib, and resources to the working directory
when used in a multiuser environment such as galaxy.