samtools 0.1.19: update permissions for vcfutils.pl
In #3900 we run into failures during the mulled build
when unpacking the samtools 0.1.19 package. It complains about
permissions when installing `vcfutils.pl`:

https://travis-ci.org/bioconda/bioconda-recipes/jobs/209609560#L1767
```
[Mar 10 03:48:15] DEBU Packing failed due to open build/dist/bin/vcfutils.pl: permission denied
[Mar 10 03:48:15] WARN Local execution errorred with error [open build/dist/bin/vcfutils.pl: permission denied], retrying with remote execution
```
The cause appears to be that `vcfutils.pl` does not have read
and execute permissions set for non-root users:
```
/anaconda/pkgs/samtools-0.1.19-1/bin:
total 2.4M
-rwxr-xr-x 1 root root 643K Apr 20  2016 bcftools
-rwxr-xr-x 1 root root 1.8M Apr 20  2016 samtools
-rwxrwx--- 1 root root  16K Apr 20  2016 vcfutils.pl
```
This updates the samtools 0.1.19 recipe to fix permissions on vcfutils.pl,
requiring some modernization of the recipe to build.