Consider env matrix updates only when on a branch called "bulk/*".
Together with PR #96, this allows us to perform bulk updates in an iterative way,
without using blacklist. The workflow is as follows:

1. Create a PR with an updated env matrix, with a target branch called "bulk/*".
2. The PR will be tested, but bioconda-utils will only build recipes that have changed in addition to the env matrix.
3. Once the PR is merged, bioconda-utils will build the changes induced by the new env matrix. Whenever a recipe has been successfully built, the package will be immediately uploaded to anaconda and biocontainers (mulled).
4. If errors occur, they can be fixed in a PR against the bulk branch. In such a case, we start again at step 1. Importantly, since successfully built packages have been already uploaded, we won't do redundant work and the list of packages will become shorter with each iteration.

If a timeout occurs during above process, once can simply restart the build job of the bulk branch, and bioconda-utils will go on with the still missing recipes.

For these changes to work, PR #96 has to be merged first.