RNFtools 0.3.1.0 (#5012)
* RNFtools 0.3.1.0

* Install reportlab

* Add cairosvg

* IQTree: new version (#5011)

* iqtree 1.5.5

* previous version of iqtree: 1.5.3

* correct hash

* Bump IgDiscover to 0.8.0 (#5008)

* Update of the Blast-legacy/2.2.22 recipe (#5009)

* Blast-legacy 2.2.22 - meta.yaml - Fix download url

Update of the download link of the Blast-legacy 2.2.22 precompiled binaries that was no longer valid.
Switch build number to 1.

* Blast-legacy 2.2.22 - build.sh - BLASTMAT environment variable

The activation of the conda package should now define a BLASTMAT environement variable pointing the the "data" sub-folder of the Blast installation folder.
On the opposite, the deactivation should unset this environment variable.

* Blast-legacy 2.2.22 - meta.yaml - BLASTMAT env variable display

The value of the BLASTMAT environment variable is now displayed during the tests.

* add new fancy test

* Update meta.yaml

* Snakemake 3.13.2 (#5014)

* Snakemake 3.13.2

* Don't use noarch.

Using noarch also implies independence from the Python version. Snakemake
needs Python >=3.3 though. Further, it has a command line interface, and conda
hardcodes the Python version into that.

* epic 0.1.25 (#5015)

* Remove some unnecessary dependencies