downloader R/CRAN package (#5389)
* Downloader R/CRAN package

Provides a wrapper for the download.file function, making it possible
to download files over HTTPS on Windows, Mac OS X, and other Unix-like
platforms. The 'RCurl' package provides this functionality (and much
more) but can be difficult to install because it must be compiled with
external dependencies. This package has no external dependencies, so it
is much easier to install.

* downloader R/CRAN package

fixed single quotes under the summary sextion