Add version requirement for openjdk
The version requirement forces openjdk to come from the
conda-forge channel and not the default channel.

This is discussed extensively in the pull request thread,
underlying problem was with openjdk coming from wrong
channel and no mechanism existing for specifying the
channel name if conda < version 3.

Problem originated with `track_features`
added to defaults::conda-forge (see discussion
[here](https://github.com/conda/conda/issues/6948#issuecomment-369368081)).

Solution is to specify a version of
openjdk that forces openjdk to come from
conda-forge and not default, which is
>8.0.121 (last version without `track_features`,
see discussion
[here](https://github.com/conda/conda/issues/6948#issuecomment-369360906)).