pypyFLOW v1.0.0
In the absense of an official tag on the upstream repository,
this is the commit which bumped the version number to 1.0.0

See PacificBiosciences/pypeFLOW#51

Note that the dependencies declared in the setup.py changed,
which is reflected in the meta.yaml dependencies. It now
requires networkx.

Also, the updated setup.py declares additional python libraries
which we now test can be imported.