Update BioConda pypeFLOW package to v1.0.0 (#9024)
* pypyFLOW v1.0.0

In the absense of an official tag on the upstream repository,
this is the commit which bumped the version number to 1.0.0

See PacificBiosciences/pypeFLOW#51

Note that the dependencies declared in the setup.py changed,
which is reflected in the meta.yaml dependencies. It now
requires networkx.

Also, the updated setup.py declares additional python libraries
which we now test can be imported.

* Use the merge commit for the v1.0.0 version bump.

Hat tip Renan Valieris, the version 1.0.0 bump commit
79c77235eb10e7b36a98ada891e1e6ef454aed58 was part of
a pull request which make other changes, thus merge
commit e331fad50be98d83e95a4a354bed2c81a8351392 is a
better representative of v1.0.0. See
https://github.com/PacificBiosciences/pypeFLOW/pull/39

This pull request also expanded the supported range of
networkx versions.

* Remove space in networkx dependency version definition