Update to pypeFLOW v1.1.0
In the absense of an official tag on the upstream repository,
this is the commit which merged pull request 20 into the
master branch, and that included the version bump.

See PacificBiosciences/pypeFLOW#51

There is a new futures dependency declared in their setup.py,
which is a backport for Python 2 only. Given pypeFLOW does not
yet support Python 3 this is perhaps not worth noting, but I
marked this dependency as conditional.