Specify flacon 2.1.4 wants pypeflow 2.0.2 (or later 2.0.x?)
Recent FALCON tags are to the FALCON-integrate repository,
which uses git submodules to bundle together the dependencies.
That indirectly tells us the hash of the relevant commit on
the main FALCON repository, e.g.

https://github.com/PacificBiosciences/FALCON-integrate/tree/2.1.4
--> FALCON @ c1e0d45

As of this commit, the falcon setup.py says "pypeFLOW >= 2.0.0"

And, FALCON-integrate lists some bundled depenencies, including:
--> pypeFLOW @ cfb4ace

That commit was on 2018-04-06 merging their pull request 30,
which was merged into the master with one other change as
c7222e1bfc38bb067d119a75dcbff600854c03ee on 2018-03-11 which
with a few more changes later was part of pypeFLOW v2.0.2
as of merge commit 6f8c77e262659866018c5d4ed9a06e46dfd7f010
on 2018-04-25.

Therefore the first release of pypeFLOW to include the
changes FALCON-integrate pinned was pypyFLOW v2.0.2

I have assumed pypeFLOW v2.0.2 or later 2.0.x will be fine,
to give a little flexibility in the dependency chain. It may
be it will work fine with later versions of pypeFLOW too?