Whitespace fixes [WIP] (#9085)
* HiCExplorer version 2.1.4 (#9044)

* Adding hic2cool version 0.4.1

* Update to version 2.1.4 of HiCExplorer

* Update to asqcan v0.2 (#9047)

* Update to asqcan v0.2

* Fixed md5

* [ci skip] add doi to deeptools (#8963)

* hicexplorer fix

* Update with new dependency (#9048)

* vep: bumped subversion (#9051)

* adding illumiprocessor v2.0.9 (#9045)

* geneimpacts: bumped subversion (#9052)

* r-geoaxe, r-oai, r-rbison, r-rebird, r-ridigbio, r-rvertnet, r-wicket (#8994)

* r cran packages dependencies for rspocc and rgbif, recipes generated with skeleton. Ecology tools, for integration in Galaxy

* Rpackages for ecology

* rm win tags

* Bumped star-fusion version (#9034)

* Bumped star-fusion version
* conda considers 2.6.0b<2.6

* snpsplit : new version added (#9054)

* Mmseqs2 (#9055)

* Update MMseqs2 recipe to new release 3-be8f6

* Update r-biodb to 1.1.0. (#9057)

* OrthoFinder: update to 2.2.6 (#9060)

* Orthofinder: update to 2.2.6

* fix link

* snakemake 5.1.4

* Update BioConda pypeFLOW package to v1.0.0 (#9024)

* pypyFLOW v1.0.0

In the absense of an official tag on the upstream repository,
this is the commit which bumped the version number to 1.0.0

See PacificBiosciences/pypeFLOW#51

Note that the dependencies declared in the setup.py changed,
which is reflected in the meta.yaml dependencies. It now
requires networkx.

Also, the updated setup.py declares additional python libraries
which we now test can be imported.

* Use the merge commit for the v1.0.0 version bump.

Hat tip Renan Valieris, the version 1.0.0 bump commit
79c77235eb10e7b36a98ada891e1e6ef454aed58 was part of
a pull request which make other changes, thus merge
commit e331fad50be98d83e95a4a354bed2c81a8351392 is a
better representative of v1.0.0. See
https://github.com/PacificBiosciences/pypeFLOW/pull/39

This pull request also expanded the supported range of
networkx versions.

* Remove space in networkx dependency version definition

* Bump dimspy version to 1.2.0 (#9063)

* Update to pypeFLOW v1.1.0 (#9064)

In the absense of an official tag on the upstream repository,
this is the commit which merged pull request 20 into the
master branch, and that included the version bump.

See PacificBiosciences/pypeFLOW#51

There is a new futures dependency declared in their setup.py,
which is a backport for Python 2 only. Given pypeFLOW does not
yet support Python 3 this is perhaps not worth noting, but I
marked this dependency as conditional.

* pypeFLOW 2 0 0 (#9065)

* Update for pypeFLOW v2.0.0

In the absense of an official tag on the upstream repository,
this is the commit which merged pull request 23 into the
master branch, and that included the version bump.

See https://github.com/PacificBiosciences/pypeFLOW/issues/51

* Add basic about section to pypeFLOW recipe

* Fix URL for abundancebin (#9066)

* Remove duplicated pypeFLOW about metadata (#9068)

Keeps the license clarification, not just normal BSD.

* trigger; try out logic for divergent build strings

* Update: peddy, simple_sv_annotation, bcbio (#9069)

- peddy: latest development version with support for hg38. Drops
  bioconda patch which they incorporated upstream.
- simple_sv_annotation: Fix for REJECT filter not in VCF header.
- bcbio: support for moving output and better debugging with Cromwell.

* Revert "temporarly disable uploads for bulk branch"

This reverts commit 3cf59832265de087ad00ad1975b70af125491064.

* Update for pypeFLOW v2.0.1 (#9070)

In the absence of an official tag on the upstream repository,
this is the commit which merged pull request 26 into the
master branch, and that included the version bump.

See https://github.com/PacificBiosciences/pypeFLOW/issues/51

* salmon v0.10.0

This bumps salmon to v0.10.0.  This release uses the `CONDA_BUILD=TRUE` flag during CMake configuration to force C++11 (instead of C++14) compilation.  Once conda-build 3 migration is complete, we should move this to build with C++14.

* AMPtk update to v1.2.3 (#9071)

* update to v1.2.3

* fix tab error

* Bissnp update (#9074)

* Updated to VERSION 1.0.1

* Fixed test.

* PacBio: Initial addition of `lima` (#9075)

* cyvcf2: remove plotting dependencies (see https://github.com/brentp/cyvcf2/issues/83) (#9078)

* Update for pypeFLOW v2.0.2 (#9080)

In the absence of an official tag on the upstream repository,
this is the commit which merged pull request 33 into the
master branch, and that included the version bump.

See https://github.com/PacificBiosciences/pypeFLOW/issues/51

* PacBio: Initial addition of `isoseq3` (#9076)

* rtg-tools: Update to 3.9.1 (#9077)

* PacBio: pull in lima as runtime dep for isoseq3 (#9081)

* This facilitates a simpler workflow, where all
  workflow dependencies are pulled in by just
  one recipe.

* Add rgbif and spocc recipes  (#9067)

* rgbif recipe created with conda skeleton cran

* fix recipes

* spocc recipe, build with conda skeleton cran

* enforce 2-space indentation

* enforce 2-space indentation

* first round of indentation fixes

* refactor

* recipes c-f whitespace

* recipes f-l whitespace

* recipes m-o whitespace

* recipes p-z whitespace