Salmon gcc7 (#10344)
This was a *very* long time coming, but salmon now builds under the new (cb3) compilers.  There *is* an issue with c++11/14 compatibility of the default (conda-forge) boost libraries, so we pull in `libboost` here for now to address this.  Hopefully, in the future, we can revert back to the plain `boost` deps.

* try new compiler

* commit build_config.yaml

* Update meta.yaml

Try building on OSX

* Update build_config.yaml

trying GCC/GXX even on Mac

* Rename build_config.yaml to conda_build_config.yaml

* pinning

* pinning

* boost => boost-cpp

* boost => boost-cpp

* boost => boost-cpp

* boost => boost-cpp

* boost => boost-cpp

* boost => boost-cpp

* Update conda_build_config.yaml

revert to clang on OSX.

* osx build

* osx build

* Update meta.yaml

bump llvm_version

* Update meta.yaml

version

* Update meta.yaml

version again

* Update meta.yaml

Try @bgruening's trick

* Delete conda_build_config.yaml

remove since we're not using this.

* Update meta.yaml

go back to other boost to be compatible with gcc 4.8.5

* Update build.sh

causing a problem with osx build?

* Update build.sh

huh?

* try and remove librt on OSX build

* skip linux until we have OSX

* Update meta.yaml

* Update meta.yaml

* Update build.sh

change deployment target (10.9)

* Create conda_build_config.yaml

* Update meta.yaml

ok, let's go all in on the new stuff and try to fix it from that direction

* Update meta.yaml

* Update meta.yaml

try different boost on osx?

* Update meta.yaml

this is a frustrated commit message ...

* Update meta.yaml

* Update meta.yaml

Try boost instead of libboost --- do we get the same error on OSX?

* Update meta.yaml

See if we can use boost instead of libboost on linux

* revert `boost` to `libboost` under linux.

nope ... `boost` instead of `libboost` fails under linux.

* Update meta.yaml

horrible, horrible, horrible ...

* Update meta.yaml

nope.  nope, nope, nope --- we *need* `libboost` on linux for now.  Why doesn't it work on OSX?

* Update meta.yaml

try this

* Update meta.yaml

update md5sum

* Update meta.yaml

50th time's a charm?

* Update meta.yaml

51st time's a charm ... I'm gonna guess not.

* Update meta.yaml

Ok --- last attempt *worked*!  Now, try to make everything C++14, and re-enable linux build.

* Update meta.yaml

have to update md5sum

* Update meta.yaml

try and build tagged v0.11.3