Made changes necessary to prevent tests from causing a hang during build.
`make test` was getting piped input on STDIN which is was passing to t/run_tests.t, which then passed it along to the script being tested, which tried to consume it as a file and just waited. I worked around this by editing the package being installed to make the test script close STDIN. This solved the problem, but there was another issue. The tests that run via the commands defined in the meta.yaml file were not using the version it was building, but rather the old version I had previously installed (which hung still). So I removed those test commands from the meta.yaml file. Testing happens in the build.sh and it tests the correct code. If I find that tests are needed in the meta.yaml file, I should either update my installed version so the build succeeds with the tests defined or change the calls in the meta.yaml file to use the version in the work directory.

Files checked in: recipes/vcfsamplecompare/build.sh recipes/vcfsamplecompare/meta.yaml