Add check for SystemRequirements,
SystemRequirements in the DESCRIPTION file provide dependencies
which are usually not fulfilled by the bioconductor-skeleton.py.

These usually represent external dependencies.

Take for instance the package 'rsbml', which has the SystemRequirement,
libsbml (==5.10.2). This is not fulfilled by this script, and should
throw a warning to the user that there is a recipe for libsbml is also
needed.

There are many packages which have such SystemRequirements with
Bioconductor, as shown in this gist.

    https://gist.github.com/nturaga/39185116b8681264e7405b868e00d405

If there are suggestions from the Bioconda team to standardizing the format of
SystemRequirements, I could try to get them standardized across
Bioconductor packages, so we have a better chance of fulfilling them
automatically.

I'm cc-ing @mtmorgan, and @vobencha to help answer add context if
needed.