Remove htslib/samtools/bcftools dependencies
This fixes https://github.com/pysam-developers/pysam/issues/640
which is an issue with the bioconda build of pysam.
Removing those dependencies will make pysam use the bundled
subset of htslib.
A side-effect of this is that pysam and samtools/bcftools/htslib
versions are now independent and can no be mixed at will.