samtools w/ openssl-1.1.1
Another attempt to deal with python2.7.15/openssl issues. E.g.
https://github.com/PacificBiosciences/pbbioconda/issues/85

Note: samtools-1.9 was built against opensll-1.0.2, which conflicts
with the latest python2.7.15 build.
openssl-1.1.1a and openssl-1.0.2p are not ABI-compatible, so we need to
specify that.

I do not see a way to pin the old opensll (since there is ambiguity
between 1.0.2o and 1.0.2p), so I am taking the suggestion
from kylebeauchamp and rebuilding with openssl-1.1.1. That might break
some people until they upgrade python, etc. but I do not
see a better solution without actually bumping the version of samtools.