Removing pass of shell variable to echo.
This is an attempt to address a warning from the remote circleci build that passing a variable as an argument generated from the output of the tool is a security risk.  It might still complain due to the if conditional checking the variable too.  I'll just have to see.

Files checked in: recipes/barcode_splitter/run_test.sh