Update ngless to v0.11.1
This is redistributing the binary, instead of compiling from scratch.

This is far from the ideal solution, but given that there is no
currently-working method to obtain a working recent-enough GHC on the
bioconda build systems, I do not see how else to package NGLess. I note
that stack is also packaged in this fashion.