Restore linking of PureCN analysis scripts.
The automatic updates removed the linking of the scripts. The dependencies were
pinned too hard for this formula and needed to be build on R 3.6 depite the
PureCN recipe not needing versions that were that new, so I removed the pinning
on those formulas.

I also added in checks to make sure the scripts are properly linked and
available.