Fix compilation
The configure script is non-standard and always tries to run "gcc" on Linux,
no matter which environment variables are set. Fix this by patching
various configuration files and the configuration script.

The actual aim of this commit was to get the recipe to install the .h files
necessary to actually use the library from other tools (in particular,
IgBLAST). The "make install" command doesn’t actually install any headers,
so we now just copy over the header files manually (as Debian does it).