Add kmer-db 1.6.2
kmer-db has a two possible executables depending on whether you want
AVX or AVX2 support. I used the same approach to that of RAXML to build
both executables, and make them available with the the -avx1 or -avx2
suffixes.

Fix linting issue

Fix test commands to be a list

Change suffixes avx1,avx2 to AVX,AVX2