pb-assembly: fix version for pb-falcon dependency (#18462)
* pb-assembly: fix version for pb-falcon dependency

The fix is needed in order to have the most recent pb-falcon version
installed (> 0.3.0). Otherwise, a version where fc_ovl_filter does not
have the --min-idt and --ignore-indels options is installed (2.2.1-0).
This makes it impossible to use the newest configuration files for HiFi
reads.

* pb-assembly: fix version for pb-falcon dependency

The fix is needed in order to have the most recent pb-falcon version
installed (> 0.3.0). Otherwise, a version where fc_ovl_filter does not
have the --min-idt and --ignore-indels options is installed (2.2.1-0).
This makes it impossible to use the newest configuration files for HiFi
reads.

* pb-assembly: fix version for pb-falcon dependency

The fix is needed in order to have the most recent pb-falcon version
installed (> 0.3.0). Otherwise, a version where fc_ovl_filter does not
have the --min-idt and --ignore-indels options is installed (2.2.1-0).
This makes it impossible to use the newest configuration files for HiFi
reads.