Update gemma to 0.98.1
After a lot of trial and error I managed to get the latest release of
gemma to build. This version contains a critical bug fix related to
association testing with missing values using PLINK data. I bet the
build could be done in a much more clean manner, but for now it seems to
work.