Add zlib dependency (#19370)
* Update meta.yaml to remove mpi4py

mpi4py>3.0 is only available for py37, and requires the conda-forge channel. Because it is not a hard dependency and to keep things simple I'll exclude it. Users can install it separate if they want to use it.

* Update meta.yaml

* Update meta.yaml to include samtools

samtools used to be a dependency of pysam, I think. I guess not anymore...

* add zlib requirement for vsearch to use gz files