Use unpinned llvm-openmp for build, host, and run
Do this in place of pinned openmp

At https://github.com/conda-forge/conda-forge-pinning-feedstock/blob/master/recipe/conda_build_config.yaml
you can see there is a pinned version of openmpi corresponding to the versions
of openmp referenced in #10583.

From Iqtree's compliation guide:

NOTE: Please be aware that OpenMP and OpenMPI are different

I wonder if the this was a mistake in the #10583.

Co-authored-by: Matthew Dillon <matthewrdillon@gmail.com>
Co-authored-by: Evan Bolyen <ebolyen@gmail.com>