[ci skip] Merge PR 21958
Merge PR #21958, commits were: 
 * remove zlib
 * 4.5covid19 has a different zip structure; correct cp

for the 4.5covid19 release, snpEff does not reside in a subdir within the source zip
 * include sub-recipe for snpEff 4.5covid19

A special build of snpEff has been published that comes pre-packaged with the reference for SARS-CoV-2. This adds a sub-recipe for that build of snpEff, while leaving the main recipe unchanged. The main recipe should be updated once the primary release of snpEff has been updated to v4.5.