update checksum of test
this version of insilicoseq fixed a bug, where insilicoseq crashed when
a user-made error model had a probability 0 of having sequences with
qualities above Q20, because the mean sequence quality was calculated
incorrectly. As a result of this bugfix, sequences now have marginally
higher quality, hence the need to modify the checksums for the test.