Remove macOS ._ files
Those files get integrated into the metadata fields of the file system and are then removed. This makes the verification fail on macOS. 
This is only a problem for Conda, not normal installation, therefore I would remove them here. (It may be argued that they should not be in the release file at all)