hint: use bgzip from htslib rather than legacy tabix package
HiNT uses `CMD | bgzip -c > FILE` (and the pytabix Python module),
which is better supported by the up-to-date bgzip provided by the
htslib package already pulled in by the samtools requirement.

Bump the samtools (and htslib) requirement to ensure that both use
libhts.so.3 -- the soversion was bumped between 1.9 and 1.10.