Add tabix as a dependency
whatshap depends on indexed VCF input, but does not produce and index
for the VCF files it produces. By adding tabix as a dependency for
whatshap, the same container/environment can be used to create the VCF
file with whatshap, and index it using tabix.

This is especially useful when running multiple whatshap commands in
succession (e.g. whatshap Phase and whatshap Haplotag).