[ci skip] Merge PR 27687
Merge PR #27687, commits were: 
 * Restore and update 0001-PATCH-macos-10.9.patch
 * Compiler fixes
 * Add `dev_url`, use SPDX licence tag
 * Upgrade IQ-TREE to v2.1.2

Upgrades IQ-TREE to the latest version which includes COVID-19-specific
optimizations added in version 2.1.0. For backwards compatibility, the
build script copies the new "iqtree2" binary to "iqtree".

Since the Mac-specific patch could no longer be applied to the latest
release, this commit removes that patch and references to it.