[ci skip] Merge PR 30081
Merge PR #30081, commits were: 
 * Update augur to 12.1.1

Additionally, drop Snakemake as an augur dependency. Snakemake is not
technically required for augur to run and is only part of augur's
development environment. More importantly, Snakemake's own dependency
graph has gotten so complex that the inclusion of Snakemake as an augur
dependency causes the Conda solver to exceed the amount of memory
allocated on Circle CI's Medium Linux Machine (>7.5 GB).