[ci skip] Merge PR 31162
Merge PR #31162, commits were: 
 * Merge branch 'master' into trinity_sort
 * Merge branch 'bioconda:master' into trinity_sort
 * Add coreutils as trinity dependency.

Trinity uses some features of the sort command not present in the BusyBox implementation.
In some container images based on BusyBox (such as some Biocontainers base images),
this results in many noisy log messages about incorrect invocation of sort. Installing
the GNU coreutils version of sort as a dependency solves this, and ensures parallel sorting
is available as an optimization.