pbcore/1.2.7: comment out Test Command
running unit tests is discouraged as part of the package build (as
I was told in the bioconda gitter), but I'm not outright deleting
the line because this helped in local testing to pinpoint the
dependency version constraints. The test fails on CircleCI because
the test requirements are not installed.