[ci skip] Merge PR 31786
Merge PR #31786, commits were: 
 * pbcore/1.2.7: comment out Test Command

running unit tests is discouraged as part of the package build (as
I was told in the bioconda gitter), but I'm not outright deleting
the line because this helped in local testing to pinpoint the
dependency version constraints. The test fails on CircleCI because
the test requirements are not installed.
 * pbcore/1.2.7: bump build number
 * pbcore/1.2.7: run test suite

And we can drop the sphinx test requirement.
 * pbcore/1.2.7: require pysam exactly at 0.8.4

with pysam 0.9.0, two of the unit tests fail. It's mostly compatible,
but not fully.
 * pbcore/1.2.7: use setup.py install to build

Using pip masks errors in dependency constraints because it will just retrieve
what it needs from pypi.