removed conda init script for gromacs_plumed
The conda activate script was modifying the PATH
variable to include the path to gmx, but without
any equivalent script to remove this path when
deactivating the environment.

As we have a selector script in bin/gmx which carries
out the same SIMD checking process, I've removed
the code which creates the conda activation script.