pysam: Patch to fix pysam.bcftools.mpileup() crashes
Several symbols inadvertently exist in both libcbcftools.xyz.so
and libcsamtools.xyz.so. As both libraries are loaded simultaneously,
the wrong function can sometimes be called, leading to crashes.
Work around this by patching the libcbcftools build so that it uses
different names for the conflicting symbols.

Bump the build number to 2, as a 0.18.0_1 build exists (rebuilding pysam
against libdeflate 1.10) despite not appearing on the master branch.