[ci skip] Merge PR 33805
Merge PR #33805, commits were: 
 * Update meta.yaml
 * perl-bioperl: make test-command a string

I meant to use the program 'true', but yaml interpreted it as a bool
without the quotes.
 * Avoid the solver giving broken dependencies.
 * Merge remote-tracking branch 'upstream/master' into perl-bio-searchio-hmmer
 * perl-bioperl{,-core}: clean up pre-existing lint
 * prokka: remove pinning of perl and bioperl versions

Now that Bio::SearchIO::hmmer is packaged, we can use it as part
of the latest bioperl distribution and there's no longer a need
for this pinning.

This reverts the workaround a32603c7e52e5edf69119b4847b4288fad2fc82b
in favor of resolving the underlying problem.
 * Add perl-bio-searchio-hmmer

This module was split out of the main bioperl distribution in
version 1.7.3 (https://metacpan.org/dist/BioPerl/changes), but
is still needed by Prokka (#33356).
 * perl-bioperl-core: add run-time depends on perl-db_file

This is needed for Bio::Index objects and classes that descend
from it for the creation of index files.