[ci skip] Merge PR 35235
Merge PR #35235, commits were: 
 * Update the build number for entrez-direct
 * Add wget as a dependency for entrez-direct

entrez-direct requires either curl or wget to run. wget is included in
busybox, but not in the extended base image, so it only now became missed.
Add it as a proper dependency.
 * Add extended-base requirement to entrez-drect recipe

This helps to resolve issues with SSL certificates via openssl which is one
of the use cases mentioned in
https://bioconda.github.io/contributor/troubleshooting.html#using-the-extended-image
for the extended image.