[ci skip] Merge PR 36914
Merge PR #36914, commits were: 
 * Patch collections.abc imports for Python 3.10 compatibility

These classes are no longer available in collections in 3.10.
The collections.abc module was introduced in 3.3, and atropos
supports only Python >= 3.3, so fortunately there is no need
to make this conditional.
 * Reinstate atropos build

Ensure that Cython is rerun by removing the *.c output files and also
removing PKG-INFO, the presence of which signals to setup.py that it
should skip running Cython.