Remove Hail form bioconda
Hello!

This PR removes Hail from bioconda. The Hail team gets a steady stream of support requests due to the very out of date version of Hail in bioconda. Hail is currently on 0.2.100 and this package is 0.2.61. Users can easily install the latest version of Hail through pip: `pip install hail`.

Thanks for your understanding.